fix(app): migrate to app/works/[slug]/page.tsx
